smart-office/services/5-analytics-service/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
jose
passlib[bcrypt]
python-multipart